Updated Error Messages
An addition to my commit earlier. I modified the error messages and removed even more warnings! Variables are no longer shadowing  within a different scope. Error Handling for the input tab should be complete, as well as the creation of Cepheid and RR-Lyrae objects! I got rid of the "sel" method since the error messages in the progress bar do the job. Line 68 is the only warning from our GUI class but is unavoidable.